requirements.txt
einops>=0.8.0
numpy>=2.0.0
opencv-python>=4.10.0
Pillow>=10.4.0
PyYAML>=6.0.1
torch>=2.5.0
torchvision>=0.20.0
tqdm>=4.66.0
safetensors>=0.4.0
transformers>=4.40.0
huggingface_hub>=0.20.0
